feat(v2): migrate emulator streaming player to v2
Move the streaming player UI from v1 to v2 per the v1-frozen rule.
The shared streaming store and backend stay canonical; only the
frontend surface moves.

- New v2 player view (Player/Stream.vue) porting the launch/playing/
  error state machine, session claim, iframe UI auto-hide, volume,
  save/load state, fullscreen, save-and-exit and stop/release.
- Fold streaming into the single Play action via useGameActions:
  routes to /rom/:id/stream when a container is configured, else
  EJS, else Ruffle. Streaming wins over in-browser emulation.
- Hydrate streaming config in AppLayout (v2 users never mount v1
  Main.vue, where v1 put the fetchConfig call).
- Make the rom/:rom/stream route v2-only (v1 default falls back to
  Home).
- Add stream-* i18n keys to all 18 locales.
- Remove the v1 Stream player, PlayBtn cast button and Main.vue
  fetchConfig wiring.

Generated-By: PostHog Code
Task-Id: e091168f-69fd-4d4a-9f12-d473dd3f17c9
