+++ frontend/src/layouts/Main.vue
@@ -29,13 +29,11 @@ import TryV2Banner from "@/components/common/TryV2Banner.vue";
 import storeCollections from "@/stores/collections";
 import storeNavigation from "@/stores/navigation";
 import storePlatforms from "@/stores/platforms";
-import { useStreamingStore } from "@/stores/streaming";
 import type { Events } from "@/types/emitter";
 
 const navigationStore = storeNavigation();
 const platformsStore = storePlatforms();
 const collectionsStore = storeCollections();
-const streamingStore = useStreamingStore();
 
 const emitter = inject<Emitter<Events>>("emitter");
 emitter?.on("refreshDrawer", async () => {
@@ -68,7 +66,6 @@ onBeforeMount(async () => {
   if (showVirtualCollections) {
     collectionsStore.fetchVirtualCollections(virtualCollectionTypeRef.value);
   }
-  streamingStore.fetchConfig();
 
   navigationStore.reset();
 });
