+++ frontend/src/locales/cs_CZ/play.json
@@ -27,5 +27,28 @@
   "select-background-color": "Vybrat barvu pozadí",
   "select-save": "Vybrat uloženou pozici",
   "select-state": "Vybrat stav",
-  "start-fresh-hint": "Vyberte níže pro pokračování, nebo stiskněte Hrát pro nový začátek."
+  "start-fresh-hint": "Vyberte níže pro pokračování, nebo stiskněte Hrát pro nový začátek.",
+  "stream-error-generic": "An unexpected error occurred.",
+  "stream-error-load-rom": "Could not load ROM details.",
+  "stream-error-not-configured": "No streaming container is configured for {platform}.",
+  "stream-exit-fullscreen": "Exit fullscreen",
+  "stream-frame-title": "Game stream",
+  "stream-fullscreen": "Fullscreen",
+  "stream-load-autosave": "Load autosave",
+  "stream-load-state": "Load state",
+  "stream-mute": "Mute",
+  "stream-occupied-body": "{rom} has been playing since {time}.",
+  "stream-occupied-fallback": "Someone else is currently playing. Try again later.",
+  "stream-occupied-title": "Session in use",
+  "stream-save-and-exit": "Save and exit",
+  "stream-save-slot": "Save slot",
+  "stream-save-state": "Save state",
+  "stream-save-unconfirmed": "The emulator did not confirm the save. Recent progress may be lost.",
+  "stream-slot-n": "Slot {n}",
+  "stream-stop": "Stop",
+  "stream-subtitle": "Streaming",
+  "stream-try-again": "Try Again",
+  "stream-unknown-game": "Unknown Game",
+  "stream-unmute": "Unmute",
+  "stream-volume": "Volume"
 }
