+++ frontend/src/v2/router/routes.ts
@@ -37,6 +37,7 @@ export const v2RouteComponents: Partial<Record<string, V2Route>> = {
   // Wave 5 — Players
   emulatorjs: () => import("@/v2/views/Player/EmulatorJS.vue"),
   ruffle: () => import("@/v2/views/Player/Ruffle.vue"),
+  stream: () => import("@/v2/views/Player/Stream.vue"),
   // Wave 6 — Library Tools (Scan / Upload) + Pair
   scan: () => import("@/v2/views/Scan.vue"),
   upload: () => import("@/v2/views/Upload.vue"),
